Add SDP and major corrections in HMM
Adjust factor models to use only T_max, reduce series only for HMM not for simulation.
Set HMM to use prior mean and covariance.
Change the name of HMMData to MKData, this data it is not necessary generate with HMM.
Change the HMM to use z_t to estimate r_t not z_{t-1}.
Redo equation rho to use z_ {t + 1}.
Conditional initial probability with z_1 = 0.
Remove the offset for simulate function.